do some settings and refactor
