+++ dev/index.html
@@ -4,10 +4,11 @@
   <head>
     <meta charset="utf-8">
     <title>&lt;markdown-editable-components> Demo</title>
-    <script>
-      exports = {}; // TODO
-    </script>
-    <script src="../../index.js"></script>
+    <script type="module" src="../build/index.js"></script>
+    <link
+      href="https://fonts.googleapis.com/icon?family=Material+Icons"
+      rel="stylesheet"
+    />
     <!--script src="https://cdnjs.cloudflare.com/ajax/libs/markdown-it/12.0.4/markdown-it.min.js" integrity="sha512-0DkA2RqFvfXBVeti0R1l0E8oMkmY0X+bAA2i02Ld8xhpjpvqORUcE/UBe+0KOPzi5iNah0aBpW6uaNNrqCk73Q==" crossorigin="anonymous"></script-->
     <script>
       document.addEventListener("DOMContentLoaded", function() {
