+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "markdown-editable-components",
-  "version": "0.1.11",
+  "version": "0.2.0",
   "description": "Markdown web components",
   "main": "index.js",
   "module": "index.js",
@@ -49,8 +49,8 @@
     "@rollup/plugin-replace": "^2.3.1",
     "@types/chai": "^4.2.11",
     "@types/mocha": "^7.0.2",
-    "@typescript-eslint/eslint-plugin": "^2.27.0",
-    "@typescript-eslint/parser": "^2.27.0",
+    "@typescript-eslint/eslint-plugin": "^4.23.0",
+    "@typescript-eslint/parser": "^4.23.0",
     "chai": "^4.2.0",
     "deepmerge": "^4.2.2",
     "es-dev-server": "^1.46.1",
