+++ rollup.config.js
@@ -2,14 +2,13 @@ import { nodeResolve } from '@rollup/plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
 
 export default {
-    input: 'build/index.js',
-    output: {
-      file: 'index.js',
-      format: 'cjs'
-    },
-    plugins: [
-      nodeResolve(),
-      commonjs({
-      })
-    ]
-  };
+  input: 'out-tsc/src/index.js',
+  output: {
+    file: 'build/index.js',
+    format: 'cjs'
+  },
+  plugins: [
+    nodeResolve(),
+    commonjs({})
+  ]
+};
