+++ src/index.ts
@@ -1,2 +1,1 @@
 export * from './markdown-components';
-export * from './marked-renderer';
