+++ src/markdown-components/abstract/block-element.ts
@@ -1,0 +1,5 @@
+import { MarkdownLitElement } from './markdown-lit-element';
+
+export abstract class BlockElement extends MarkdownLitElement {
+  
+}
