+++ src/markdown-components/abstract/container-element.ts
@@ -1,0 +1,5 @@
+import { MarkdownLitElement } from "./markdown-lit-element";
+
+export abstract class ContainerElement extends MarkdownLitElement {
+
+}
