+++ src/markdown-components/interfaces/markdown-element.ts
@@ -1,0 +1,3 @@
+export interface MarkdownElement {
+  getMarkdown(): string
+}
