+++ src/markdown-components/markdown-break.ts
@@ -1,0 +1,15 @@
+import { html, customElement } from 'lit-element';
+import { LeafElement } from './abstract/leaf-element';
+
+@customElement('markdown-break')
+export class ThematicBreak extends LeafElement {
+  render() {
+    return html`
+      <hr />
+    `;
+  }
+
+  getMarkdown(): string {
+    return "-----------------------\n";
+  }
+}
